Introduce Mainview mixin and remove arg form EntityActions.flagResourceNotFound() method
