+++ app/View/Show.js
@@ -2,6 +2,8 @@ import React from 'react';
 import Inflector from 'inflected';
 import { shouldComponentUpdate } from 'react-immutable-render-mixin';
 
+import { hasEntityAndView, getView } from '../Mixins/MainView';
+
 import Notification from '../Services/Notification';
 import NotFoundView from './NotFound';
 
@@ -15,7 +17,11 @@ class ShowView extends React.Component {
     constructor() {
         super();
 
+        this.viewName = 'ShowView';
+
         this.shouldComponentUpdate = shouldComponentUpdate.bind(this);
+        this.hasEntityAndView = hasEntityAndView.bind(this);
+        this.getView = getView.bind(this);
     }
 
     componentDidMount() {
@@ -45,22 +51,6 @@ class ShowView extends React.Component {
         EntityStore.removeFailureListener(this.boundedOnFailure);
     }
 
-    hasEntityAndView(entityName) {
-        try {
-            const view = this.getView(entityName);
-
-            return view.enabled;
-        } catch (e) {
-            return false;
-        }
-    }
-
-    getView(entityName) {
-        entityName = entityName || this.context.router.getCurrentParams().entity;
-
-        return this.props.configuration.getEntity(entityName).showView();
-    }
-
     onChange() {
         this.setState(EntityStore.getState());
     }
@@ -74,7 +64,7 @@ class ShowView extends React.Component {
 
     onLoadFailure(response) {
         if (response.status && 404 === response.status) {
-            EntityActions.flagResourceNotFound(true);
+            EntityActions.flagResourceNotFound();
 
             return;
         }
